+++ lib/screens/common/cartIcon.dart
@@ -1,0 +1,88 @@
+import 'package:flutter/material.dart';
+import 'package:mvp/constants/themeColours.dart';
+import 'package:mvp/models/newCart.dart';
+import 'package:mvp/screens/shoppingCart/shoppingCartNew.dart';
+import 'package:provider/provider.dart';
+
+class CartIcon extends StatefulWidget {
+  @override
+  _CartIconState createState() => _CartIconState();
+}
+
+class _CartIconState extends State<CartIcon> {
+
+  // animated Route to Shoppingcart screen
+  Route _createRoute(Widget page) {
+    return PageRouteBuilder(
+      pageBuilder: (context, animation, secondaryAnimation) => page,
+      transitionsBuilder: (
+        BuildContext context,
+        Animation<double> animation,
+        Animation<double> secondaryAnimation,
+        Widget child,
+      ) =>
+          ScaleTransition(
+        scale: Tween<double>(
+          begin: 0.0,
+          end: 1.0,
+        ).animate(
+          CurvedAnimation(
+            parent: animation,
+            curve: Curves.fastOutSlowIn,
+          ),
+        ),
+        child: child,
+      ),
+    );
+  }
+  
+  // This function checks whether there is any item
+// in the cart or not
+  Widget _checkCartItems() {
+    return Container(
+      width: MediaQuery.of(context).size.width * 0.1,
+      height: 22.0,
+      decoration: BoxDecoration(
+        color: ThemeColoursSeva().pallete1,
+        shape: BoxShape.circle,
+      ),
+      child: Consumer<NewCartModel>(
+        builder: (context, cart, child) {
+          return Center(
+            child: Text(
+              cart.totalItems == null ? '0' : cart.totalItems.toString(),
+              style: TextStyle(color: Colors.white),
+            ),
+          );
+        },
+      ),
+    );
+  }
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        Padding(
+          padding: const EdgeInsets.symmetric(vertical: 6.0, horizontal: 10.0),
+          child: IconButton(
+              color: ThemeColoursSeva().black,
+              iconSize: 30.0,
+              icon: Icon(Icons.shopping_basket),
+              onPressed: () {
+                // Handle shopping cart
+                Navigator.of(context).push(_createRoute(ShoppingCartNew()));
+              }),
+        ),
+        Positioned(
+          left: 28.0,
+          top: 10.0,
+          child: _checkCartItems(),
+        ),
+      ],
+    );
+  }
+}
+
+
+
+
